requirements.txt
fastapi
uvicorn[standard]
transformers
torch
chromadb
python-dotenv
openai